+++ lightblue-ldap-crud/src/main/java/com/redhat/lightblue/crud/ldap/translator/TranslatorFromJson.java
@@ -16,7 +16,7 @@
  You should have received a copy of the GNU General Public License
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.redhat.lightblue.crud.ldap;
+package com.redhat.lightblue.crud.ldap.translator;
 
 import java.util.ArrayList;
 import java.util.List;
